Add new Login page separate from setting page Add rederect to loginPage if not authentificated else go to homePage Add authUtils that look if the user is  authentificated. Add rederection to HomePage when login sucessfully
