+++ src/App.tsx
@@ -12,6 +12,9 @@ import { AlertProvider } from "./Utils/AlertContext";
 import { SessionProvider } from "./Utils/SessionContext";
 import LabelPage from "./Pages/LabelPage/LabelPage";
 import AlertBanner from "./Components/AlertBanner/AlertBanner";
+import LoginPage from "./Pages/LoginPage/LoginPage";
+import { isAuthenticated } from "./Utils/Auth/AuthUtil";
+
 
 function App() {
   useEffect(() => {
@@ -30,11 +33,12 @@ function App() {
           </StrictMode>
           <Routes>
             <Route path="/">
-              <Route index element={<HomePage />} />
+              <Route index element={isAuthenticated() ? <HomePage /> : <LoginPage/>} />
               <Route path="Settings" element={<SettingPage />} />
               <Route path="Saved" element={<SavedListPage />} />
               <Route path="Label/:labelId" element={<LabelPage />} />
               <Route path="*" element={<NoPage />} />
+              <Route path="Login" element={<LoginPage />} />
             </Route>
           </Routes>
         </BrowserRouter>
