+++ src/Pages/SettingPage/SettingPage.css
@@ -1,3 +1,4 @@
+/* SettingPage.css */
 .settings {
   margin: 20px;
   padding: 20px;
@@ -23,52 +24,6 @@
   border-radius: 5px;
 }
 
-#login,
-#uname,
-#password {
-  display: flex;
-  flex-direction: column;
-  align-content: start;
-  justify-content: start;
-}
-
-#uname label,
-#password label {
-  width: fit-content;
-}
-#uname input,
-#password input {
-  width: 20%;
-  height: 1em;
-  padding: 10px;
-  font-size: inherit;
-  min-height: initial;
-  text-align: left;
-}
-
-#validate-uname {
-  display: flex;
-}
-#validate-uname input {
-  margin: auto 20px auto 0;
-  width: fit-content;
-}
-#validate-uname p {
-  width: 35%;
-  text-align: left;
-}
 .settings hr {
   width: 100%;
-}
-
-.send-uname {
-  width: fit-content;
-}
-
-.send-uname:disabled {
-  background-color: #666666 !important;
-  cursor: not-allowed;
-}
-.send-uname:disabled:hover {
-  transform: none !important;
 }
